requirements.txt
Adafruit-PureIO==1.1.11
Adafruit_GPIO==1.0.3
Adafruit_SSD1306==1.6.2
pillow==10.2.0
